add base class for single agent planner
